requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.9.44
atomicwrites==1.4.0
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
execnet==1.9.0
filelock==3.3.0
idna==3.2
importlib-metadata==4.5.0
iniconfig==1.1.1
jsonpath==0.82
Moka==0.0.5
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
pbr==5.6.0
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.5
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
pytest_async==0.1.1
PyYAML==5.4.1
regex==2021.10.8
requests==2.26.0
requests-toolbelt==0.9.1
selenium==3.141.0
Shelves==0.3.8
simplejson==3.17.5
six==1.16.0
stevedore==3.4.0
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.5
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.2
zipp==3.4.1
